@@ -21,27 +21,27 @@
     "rc-menu": "^9.11.1",
     "rc-slider": "^9.7.5",
     "re-resizable": "^6.9.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-collapsible": "^2.10.0",
     "react-color": "^2.19.3",
     "react-copy-to-clipboard": "^5.0.4",
     "react-datepicker": "^4.16.0",
     "react-device-detect": "^2.2.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-ga4": "^2.0.0",
     "react-highlight-words": "^0.20.0",
     "react-icons": "^4.8.0",
     "react-masonry-css": "^1.0.16",
+    "react-movable": "^3.2.0",
     "react-responsive-modal": "^6.4.2",
     "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
     "react-select": "^5.7.4",
     "react-slick": "^0.29.0",
-    "react-sortable-hoc": "^2.0.0",
     "react-switch": "^7.0.0",
     "react-table": "^7.8.0",
     "react-table-sticky": "^1.1.3",
-    "react-tabs": "^4.3.0",
+    "react-tabs": "^6.0.2",
     "react-tooltip": "^5.20.0",
     "react-transition-group": "^4.4.5",
     "react-virtualized": "^9.22.5",
