@@ -32,7 +32,7 @@ class Identify extends Component {
     });
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     this.refreshLayers(nextProps);
   }
 
