@@ -54,22 +54,22 @@ const defaultStyles = {
 
 const Sidebar = ({
   children,
-  styles,
+  styles = {},
   rootClassName,
   sidebarClassName,
   contentClassName,
   overlayClassName,
   sidebar,
-  docked,
-  open,
-  transitions,
-  touch,
-  touchHandleWidth,
-  pullRight,
-  shadow,
-  dragToggleDistance,
-  onSetOpen,
-  defaultSidebarWidth,
+  docked = false,
+  open = false,
+  transitions =true,
+  touch = true,
+  touchHandleWidth = 20,
+  pullRight = false,
+  shadow = true,
+  dragToggleDistance = 30,
+  onSetOpen = () => {},
+  defaultSidebarWidth = 0,
   rootId,
   sidebarId,
   contentId,
@@ -326,20 +326,6 @@ Sidebar.propTypes = {
   sidebarId: PropTypes.string,
   contentId: PropTypes.string,
   overlayId: PropTypes.string,
-};
-
-Sidebar.defaultProps = {
-  docked: false,
-  open: false,
-  transitions: true,
-  touch: true,
-  touchHandleWidth: 20,
-  pullRight: false,
-  shadow: true,
-  dragToggleDistance: 30,
-  onSetOpen: () => {},
-  styles: {},
-  defaultSidebarWidth: 0,
 };
 
 export default Sidebar;
